Remove legacy refreshWithoutAsync() method for HDInsightRootModule
