+++ Utils/azure-explorer-common/src/com/microsoft/tooling/msservices/serviceexplorer/RefreshableNode.java
@@ -25,12 +25,10 @@ import com.google.common.util.concurrent.FutureCallback;
 import com.google.common.util.concurrent.Futures;
 import com.google.common.util.concurrent.ListenableFuture;
 import com.google.common.util.concurrent.SettableFuture;
-import com.microsoft.tooling.msservices.components.DefaultLoader;
 import com.microsoft.azuretools.azurecommons.helpers.AzureCmdException;
 import com.microsoft.azuretools.core.mvp.ui.base.NodeContent;
+import com.microsoft.tooling.msservices.components.DefaultLoader;
 
-import javax.swing.*;
-import javax.swing.tree.TreePath;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -103,9 +101,6 @@ public abstract class RefreshableNode extends Node {
         }
     }
 
-    public void refreshWithoutAsync() {
-    }
-
     protected void refreshFromAzure() throws Exception {
     }
 
@@ -136,16 +131,26 @@ public abstract class RefreshableNode extends Node {
                             Futures.addCallback(future, new FutureCallback<List<Node>>() {
                                 @Override
                                 public void onSuccess(List<Node> nodes) {
-                                    updateName(nodeName, null);
-                                    updateNodeNameAfterLoading();
-                                    expandNodeAfterLoading();
+                                    DefaultLoader.getIdeHelper().invokeAndWait(new Runnable() {
+                                        @Override
+                                        public void run() {
+                                            updateName(nodeName, null);
+                                            updateNodeNameAfterLoading();
+                                            expandNodeAfterLoading();
+                                        }
+                                    });
                                 }
 
                                 @Override
                                 public void onFailure(Throwable throwable) {
-                                    updateName(nodeName, throwable);
-                                    updateNodeNameAfterLoading();
-                                    expandNodeAfterLoading();
+                                    DefaultLoader.getIdeHelper().invokeAndWait(new Runnable() {
+                                        @Override
+                                        public void run() {
+                                            updateName(nodeName, throwable);
+                                            updateNodeNameAfterLoading();
+                                            expandNodeAfterLoading();
+                                        }
+                                    });
                                 }
                             });
                             node.refreshItems(future, forceRefresh);
@@ -153,21 +158,16 @@ public abstract class RefreshableNode extends Node {
                     }
 
                     private void updateName(String name, final Throwable throwable) {
-                        DefaultLoader.getIdeHelper().invokeAndWait(new Runnable() {
-                            @Override
-                            public void run() {
-                                node.setName(name);
+                        node.setName(name);
 
-                                if (throwable != null) {
-                                    DefaultLoader.getUIHelper().showException("An error occurred while attempting " +
-                                                    "to load " + node.getName() + ".",
-                                            throwable,
-                                            "MS Azure Explorer - Error Loading " + node.getName(),
-                                            false,
-                                            true);
-                                }
-                            }
-                        });
+                        if (throwable != null) {
+                            DefaultLoader.getUIHelper().showException("An error occurred while attempting " +
+                                            "to load " + node.getName() + ".",
+                                    throwable,
+                                    "MS Azure Explorer - Error Loading " + node.getName(),
+                                    false,
+                                    true);
+                        }
                     }
                 }
         );
