+++ Utils/hdinsight-node-common/src/com/microsoft/azure/hdinsight/serverexplore/HDInsightRootModuleImpl.java
@@ -53,27 +53,7 @@ public class HDInsightRootModuleImpl extends HDInsightRootModule {
 
     @Override
     protected void refreshItems() throws AzureCmdException {
-        synchronized (this) { //todo???
-//            TelemetryManager.postEvent(TelemetryCommon.HDInsightExplorerHDInsightNodeExpand, null, null);
-
-            clusterDetailList = ClusterManagerEx.getInstance().getClusterDetails().stream()
-                    .filter(ClusterManagerEx.getInstance().getHDInsightClusterFilterPredicate())
-                    .collect(Collectors.toList());
-
-            if (clusterDetailList != null) {
-                for (IClusterDetail clusterDetail : clusterDetailList) {
-                    addChildNode(new ClusterNode(this, clusterDetail));
-                }
-            }
-        }
-    }
-
-    @Override
-    public void refreshWithoutAsync() {
         synchronized (this) {
-            removeAllChildNodes();
-
-            // refresh the cluster list with invalidating the cache
             clusterDetailList = ClusterManagerEx.getInstance().getClusterDetails().stream()
                     .filter(ClusterManagerEx.getInstance().getHDInsightClusterFilterPredicate())
                     .collect(Collectors.toList());
@@ -83,14 +63,6 @@ public class HDInsightRootModuleImpl extends HDInsightRootModule {
                     addChildNode(new ClusterNode(this, clusterDetail));
                 }
             }
-        }
-
-    }
-
-    @Override
-    protected void onNodeClick(NodeActionEvent e) {
-        if (!initialized) {
-            this.load(false);
         }
     }
 }
